Fixed notifications and started take insulin page.
The notifications weren't firing after the phone
had been reset. Needed to add the recieve boot
permission to the app.

Last notification of the day just continually
fired and wouldn't stop. It was because technically
setting the first notification of the day without
adding a day to the target was aiming for the past.
Which meant it should launch instantly and try to
set up the next one and it got caught up in a loop.

Created XML file for take insulin page.