Fixed insulin recommendations/notifications.
Fixed the bug preventing the current
recommendation from displaying when the take
insulin activity is launched.

Added the feature to take the user directly to the
take insulin screen while preserving a correctly
ordered back stack when they tap on the
notification.

Removed test notification when setting up next
notification.